application/requirements.txt
fastapi[all]==0.104.0
pydantic==2.4.2
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==2.0.21
uvicorn==0.23.2
psycopg2-binary==2.9.9
pyjwt[crypto]==2.8.0
fastapi-cache2[redis]==0.2.1
celery==5.3.6
flower==2.0.1
pillow==10.2.0
sqladmin==0.16.1
gunicorn==21.2.0
alembic==1.13.1
asyncpg==0.27.0
aioredis==2.0.1
passlib==1.7.4
aiofiles==23.2.1